+++ NEWS.md
@@ -1,3 +1,8 @@
+# bnma 0.3
+
+* Modified treatment.effect function so that it has option to specify global mean and sd. This might be useful when using multiple imputation.
+* Added mcmc.add function which conveniently combines mcmc results. This would be useful when performing analysis using multiple imputation.
+
 # bnma 0.2
 
 * Extended the package to include IPD network meta-analysis model